v1.2

+++ src/main/java/crud/dao/UserDao.java
@@ -11,10 +11,12 @@ public interface UserDao {
 
     void updateUser(User user);
 
-    void removeUser(int id);
+    void removeUser(Long id);
 
-    User getUserById(int id);
+    User getUserById(Long id);
 
     List<User> listUser();
+
+
 
 }
